requirements.txt
google-adk
litellm
google-generativeai
python-dotenv
langchain
langchain-community
python-multipart
youtube_transcript_api